requirements.txt
click==6.6
path.py==8.2.1
lxml==3.6.4
git+https://github.com/liamks/pyitunes.git